Seperate GUI from command line
