+++ fahbench/cmd/cmd-main.cpp
@@ -6,18 +6,16 @@
 #include <boost/function.hpp>
 #include <boost/format.hpp>
 
-#include "FAHBenchVersion.h"
-#include "SimulationWorker.h"
+#include "../FAHBenchVersion.h"
+#include "../Simulation.h"
+#include "../GPUInfo.h"
+#include "../Utils.h"
 #include "CommandLineUpdater.h"
-#include "GPUInfo.h"
-#include "Utils.h"
 
 namespace po = boost::program_options;
 namespace mm = OpenMM;
 using std::string;
 using std::map;
-
-const static string openmm_data_dir = "share/openmm_data/";
 
 string getGpuDesc() {
     std::stringstream ss;
@@ -25,17 +23,17 @@ string getGpuDesc() {
     try {
         // OpenCL
         map<string, OpenCLId> opencl_devices = GPUInfo::getOpenCLDevices();
-        for (auto const & kv : opencl_devices){
+        for (auto const & kv : opencl_devices) {
             ss << boost::format("OpenCL Device:\t%1%\tDevice id: %3%\tPlatform id: %2%") % kv.first % kv.second.platformId % kv.second.deviceId << std::endl;
         }
     } catch(const std::exception & err) {
         ss << err.what() << std::endl;
     }
 
-    try{
+    try {
         // Cuda
         map<string, int> cuda_devices = GPUInfo::getCUDADevices();
-        for (auto const & kv : cuda_devices){
+        for (auto const & kv : cuda_devices) {
             ss << boost::format("CUDA Device:\t%1%\tDevice id: %2%") % kv.first % kv.second << std::endl;
         }
     } catch(const std::exception & err) {
@@ -50,7 +48,7 @@ int main(int argc, char **argv) {
     auto setSys = std::bind(&Simulation::setSysFile, &simulation, std::placeholders::_1);
     auto setInt = std::bind(&Simulation::setIntFile, &simulation, std::placeholders::_1);
     auto setState = std::bind(&Simulation::setStateFile, &simulation, std::placeholders::_1);
-    
+
     po::options_description desc("FAHBench options");
     desc.add_options()
     ("help,h", "produce help message")
@@ -67,7 +65,7 @@ int main(int argc, char **argv) {
     ("disable-accuracy-check", "Don't check against the reference platform")
     ("enable-accuracy-check", "Check against the reference platform (default)")
     ("nan-check", po::value<int>(&simulation.nan_check_freq)->default_value(0),
-                  "Frequency to perform NaN checks during benchmark. 0 - disable.")
+     "Frequency to perform NaN checks during benchmark. 0 - disable.")
     ("device-info,d", "List GPU device information")
     ;
 
@@ -86,13 +84,13 @@ int main(int argc, char **argv) {
         return 1;
     }
 
-    if (vm.count("version")){
+    if (vm.count("version")) {
         std::cout << "FAHBench version " << getVersion() << std::endl;
         std::cout << "OpenMM version " << mm::Platform::getOpenMMVersion() << std::endl;
         return 1;
     }
 
-    if (vm.count("device-info")){
+    if (vm.count("device-info")) {
         std::cout << "Available devices" << std::endl;
         std::cout << "-----------------" << std::endl << std::endl;
         std::cout << getGpuDesc() << std::endl;
@@ -102,45 +100,13 @@ int main(int argc, char **argv) {
     if (vm.count("disable-accuracy-check")) {
         simulation.verifyAccuracy = false;
     }
-    
-    if (vm.count("enable_accuracy-check")){
+
+    if (vm.count("enable_accuracy-check")) {
         simulation.verifyAccuracy = true;
     }
-    
+
     std::cout << simulation.summary() << std::endl;
-    
+
     CommandLineUpdater updater;
     simulation.run(updater);
-    
-    /*
-    
-    simulation.configure();
-
-    SimulationWorker sWorker;
-    if(simulation.sysFile != "" && simulation.stateFile != "") {
-        std::cout << "Custom run: " << std::endl;
-        std::cout << simulation.sysFile << " " << simulation.stateFile << std::endl;
-        sWorker.startSimulation(simulation);
-    } else {
-        if(vm["solvent"].as<string>() == "explicit") {
-            std::cout << "Explicit: " << std::endl; // TODO: Move logic
-            std::cout << simulation.numSteps << " steps" << std::endl;
-            simulation.sysFile = openmm_data_dir + "dhfr.system.explicit.xml";
-            simulation.stateFile = openmm_data_dir + "dhfr.state.explicit.xml";
-            simulation.integratorFile = openmm_data_dir + "dhfr.integrator.explicit.xml";
-            sWorker.startSimulation(simulation);
-        }
-        if(vm["solvent"].as<string>() == "implicit") {
-            std::cout << "Implicit: " << std::endl;
-            simulation.numSteps = NUMSTEPSIMPLICIT; // TODO: move this logic
-            simulation.sysFile = openmm_data_dir + "dhfr.system.implicit.xml";
-            simulation.stateFile = openmm_data_dir + "dhfr.state.implicit.xml";
-            simulation.integratorFile = openmm_data_dir + "dhfr.integrator.implicit.xml";
-            sWorker.startSimulation(simulation);
-
-        }
-    }
-    
-    */
-
 }
